feat(Settings): add copilot and gemini client settings
improve settings scripts structure
